issue-027 Reorganize util submodule
