+++ AstroPhotography/core/ApProcess.py
@@ -1602,13 +1602,13 @@ class ApProcess:
             match.
             If set to ``None`` then the first navigated file is used.
             (resample_images_to_match parameter)
-        resampled_file_prefix : str. optional, default='resampled_'
+        resampled_file_prefix : str. optional, default="resampled_"
             The part of the resampled image file name in front of the
             filter name. For example, for a ``Green`` filter output
             image name of ``M101_Supernova_2023ixf_Green_resamp_weighted.fits``
             then ``resampled_file_prefix='M101_Supernova_2023ixf_'``.
             (resample_images_to_match parameter)
-        resampled_file_suffix : str, optional, default='_resamp_weighted.fits'
+        resampled_file_suffix : str, optional, default="_resamp_weighted.fits"
             The part of the resampled image file name after the filter
             name. **Note** that this should include the file suffix, which
             should include ``.fits`` (or further, e.g. ``.fits.gz``)
@@ -2500,6 +2500,7 @@ class ApProcess:
         resampled_summary_plot: str | None = None,
         composite_summary_plot: str | None = None,
         combine_type: str = "WEIGHTED",
+        composite_plot_type: str = "stiff",
     ) -> tuple[Any, Any]:
         """
         Resample and combine all navigated images to match the WCS defined
@@ -2576,6 +2577,11 @@ class ApProcess:
             I recommened including some information on the chosen type in your
             ``resampled_file_suffix``, e.g. in the example above the suffix was
             ``resampled_file_suffix='_resamp_weighted.fits'``.
+        composite_plot_type : {"stiff", "lupton"}, optional, default="stiff"
+            Method used for generating 3-color composite images. If
+            ``"lupton"`` then the ``astropy`` method ``make_lupton_rgb`` will
+            used. If ``"stiff"`` then :class:`ApFitsRasterizer` will be used
+            which in turn uses Astromatic ``stiff``.
 
         Returns
         -------
@@ -2605,7 +2611,7 @@ class ApProcess:
         -----
 
         Astromatic ``swarp`` must be installed. The rasterized TIFF-format
-        quick look images also require that ``swarp`` is installed.
+        quick look images also require that ``stiff`` is installed.
 
         Images must have valid WCS headers in order to be resampled and
         stacked, either from external sources or generated by
@@ -3014,25 +3020,46 @@ class ApProcess:
             )
             self._logger.info(f"Resampled image summary plot generated: {resampled_summary_plot}")
         if composite_summary_plot is not None:
-            self._logger.debug(
-                "Generating a summary color composite plot using "
-                f"WCS info, tick spacing {angle_tick_spacing} arcmin,"
-                f" and swap_radec_axis={swap_axis}."
-            )
-            util.make_lupton_threecolor_plots(
-                resampled_images,
-                composite_summary_plot,
-                extnum=self._extnum,
-                usewcs=True,
-                vmin=None,
-                xaxlim=None,
-                yaxlim=None,
-                qval=qval,
-                stretchval=stretch,
-                verbose=True,
-                angle_tick_spacing_am=angle_tick_spacing,
-                swap_radec_axis=swap_axis,
-            )
+            if "lupton" in composite_plot_type:
+                self._logger.debug(
+                    "Generating a summary color composite plot using "
+                    f"WCS info, tick spacing {angle_tick_spacing} arcmin,"
+                    f" and swap_radec_axis={swap_axis}."
+                )
+                util.make_lupton_threecolor_plots(
+                    resampled_images,
+                    composite_summary_plot,
+                    extnum=self._extnum,
+                    usewcs=True,
+                    vmin=None,
+                    xaxlim=None,
+                    yaxlim=None,
+                    qval=qval,
+                    stretchval=stretch,
+                    verbose=True,
+                    angle_tick_spacing_am=angle_tick_spacing,
+                    swap_radec_axis=swap_axis,
+                )
+            elif "stiff" in composite_plot_type:
+                util.make_stiff_threecolor_plots(
+                    resampled_images,
+                    composite_summary_plot,
+                    rasterizer=self._rasterizer,
+                    extnum=self._extnum,
+                    min_cut=None,
+                    max_cut=None,
+                    min_percent=[50.0, 50.0, 50.0],
+                    max_percent=[99.9, 99.9, 99.9],
+                    negative=False,
+                    binning=2,
+                    xaxlim=None,
+                    yaxlim=None,
+                    verbose=True,
+                )
+            else:
+                err_msg = f"Unexpected composite_plot_type={composite_plot_type}"
+                self._logger.error(err_msg)
+                raise RuntimeError(err_msg)
             self._logger.info(f"Summary color composite plot generated: {composite_summary_plot}")
 
         return resampled_images, resampled_info_table
